introduce Pop type + lift pop scale to support billion-scale civs (#5)
Adds a wide `Pop` type (Q96.32, i128-backed, ±~4e28 range) to
`sim_arith` for any quantity that is a count of individuals —
`Cohort` brackets, carrying capacities, food demand, civ
aggregates, migration deltas. The existing `Real` (Q32.32, i64,
±2.1B ceiling) stays for rates, fractions, multipliers, and
everything else; mixed arithmetic ops (`Pop * Real -> Pop`,
`Pop / Pop -> Real`, `Pop / Real -> Pop`) keep the type
boundaries explicit and compiler-checked.

Then retunes the calibration so the new headroom is actually
useful:

  * Baseline `carrying_capacity_per_unit` 2,500 -> 50,000 (20x).
    Earth-equivalent stone-age cell now holds ~50k people.
  * Tier-2 capacity multipliers: BulkCultivation x2 -> x5,
    AnimalSymbiosis x1.5 -> x2.5 (agriculture is the headline
    demographic-transition jump).
  * Tier-4 industrial-era: Mechanisation x3 -> x10,
    ChemicalSynthesis x1.2 -> x3, MedicalIntervention x1.1 -> x2.
  * Tier-5 information-age: AdvancedMedicine x1.2 -> x3,
    GeneticManipulation x1.2 -> x2.
  * Full tech stack now spans ~8,000x (vs ~99x before), matching
    the real paleolithic -> modern density growth of ~1000-5000x.
  * Constructor default `carrying_capacity_per_unit` 500 ->
    50,000 in lockstep.
  * Emergent-civ founding band 20 -> 200 people so a fresh civ
    starts at ~0.4% of cell cap rather than 0.04%.

Wire format: `Pop` raw bits (i128) are serialized as JSON
decimal strings via `pop_bits_serde` / `pop_bits_vec_serde`
helpers because `serde`'s internal `Content` buffer (used by
`#[serde(tag = "kind")]` enums) has no I128 variant. String
encoding round-trips cleanly and is portable across every JSON
consumer.

Trajectory on a typical Earth-equivalent seed:

  | Era         | Tools unlocked | Per-cell cap | Single-civ pop |
  |-------------|----------------|--------------|----------------|
  | Paleolithic | none           | ~50k         | ~1-5M          |
  | Agricultural| tier-2         | ~700k        | ~20-70M        |
  | Industrial  | tier-4         | ~60M         | ~2-6B          |
  | Modern      | tier-5         | ~400M        | ~tens of B+    |

Q96.32 headroom (±~4e28) carries even speculative future-age
megacivilizations without saturation.

All workspace tests pass (`cargo test --workspace --release`).
Empirical 150-year seed-495 smoke run produces sensible
stone-age civs at ~270-800 founding pop; longer runs reach the
projected trajectories above as the tech tree fills.

+++ protocol/src/civ_events.rs
@@ -195,7 +195,17 @@ pub struct CivFounded {
     /// back to `"civ {civ_id}"`.
     #[serde(default)]
     pub name: String,
-    pub initial_population_q32: i64,
+    /// Q96.32 raw bits of the founding cohort population. Backed by
+    /// `Pop` (i128) so modern/future-age civs can carry > 2.1B
+    /// without saturating. Field kept as `_q32` for legacy NDJSON
+    /// compatibility — the Q-format suffix here documents the
+    /// fractional precision (32 bits) common to `Real` (Q32.32) and
+    /// `Pop` (Q96.32). Integer width is now i128. Serialized as a
+    /// JSON string (`serde`'s tagged-enum buffer doesn't support
+    /// `i128` natively — see `pop_bits_serde`).
+    #[serde(with = "crate::pop_bits_serde")]
+    #[schemars(with = "String")]
+    pub initial_population_q32: i128,
     pub founding_figure_count: u32,
     /// Cell ids the civ claims as its territory at founding
     /// Used by the post-run report to render per-civ
@@ -219,18 +229,24 @@ pub struct CivTerritoryChanged {
     pub tick: u64,
     pub civ_id: u32,
     pub claimed_cells: Vec<u32>,
-    /// Q32.32 raw bits of the cohort population at the time of the
+    /// Q96.32 raw bits of the cohort population at the time of the
     /// change. Lets consumers correlate territory size with the
-    /// underlying pressure.
-    pub population_q32: i64,
+    /// underlying pressure. Backed by `Pop` (i128) — modern civs
+    /// reach into the billions, well beyond Q32.32's ±2.1B ceiling.
+    /// Wire-encoded as a JSON string (see `pop_bits_serde`).
+    #[serde(with = "crate::pop_bits_serde")]
+    #[schemars(with = "String")]
+    pub population_q32: i128,
     /// Per-cell population breakdown. One entry per cell in
     /// `claimed_cells`, in the same order. Lets the post-run
     /// keyframe maps render density (cell pop / cap) as ASCII
     /// shading rather than just ownership-by-civ. Sums to
-    /// `population_q32`.
-    #[serde(default)]
-    pub cell_populations_q32: Vec<i64>,
-    /// Per-cell carrying capacity (Q32.32 raw). One entry per cell
+    /// `population_q32`. Wire-encoded as a JSON array of decimal
+    /// strings (see `pop_bits_vec_serde`).
+    #[serde(default, with = "crate::pop_bits_vec_serde")]
+    #[schemars(with = "Vec<String>")]
+    pub cell_populations_q32: Vec<i128>,
+    /// Per-cell carrying capacity (Q96.32 raw). One entry per cell
     /// in `claimed_cells`, in the same order. Computed from the
     /// civ's `cell_capacity` (tech × terrain × seasonal ×
     /// biosphere) at the moment of the territory change. Lets the
@@ -238,9 +254,11 @@ pub struct CivTerritoryChanged {
     /// fraction of *its own* cap — digit `9` = saturated, `0` =
     /// nearly empty — so density tells a "how full is this cell"
     /// story rather than absolute headcount. Empty for older event
-    /// logs; consumers fall back to a frame-relative scale.
-    #[serde(default)]
-    pub cell_capacities_q32: Vec<i64>,
+    /// logs; consumers fall back to a frame-relative scale. Wire-
+    /// encoded as a JSON array of decimal strings.
+    #[serde(default, with = "crate::pop_bits_vec_serde")]
+    #[schemars(with = "Vec<String>")]
+    pub cell_capacities_q32: Vec<i128>,
 }
 
 /// A civilization collapsed. Emitted once per civ lifecycle.
@@ -254,8 +272,11 @@ pub struct CivCollapsed {
     pub tick: u64,
     pub civ_id: u32,
     pub reason: String,
-    /// Q32.32 raw bits of the cohort population at collapse.
-    pub final_population_q32: i64,
+    /// Q96.32 raw bits of the cohort population at collapse.
+    /// Wire-encoded as a JSON string (see `pop_bits_serde`).
+    #[serde(with = "crate::pop_bits_serde")]
+    #[schemars(with = "String")]
+    pub final_population_q32: i128,
     pub final_figure_count: u32,
 }
